Move coordinator tests to test_init and simplify assertions
